feat: create loan label component
